+++ Resources/Unused/PredecessorDetailResource.cs
@@ -5,5 +5,5 @@ public class PredecessorDetailResource
   public long StepworkId { get; set; }
   public long RelatedStepworkId { get; set; }
   public long Type { get; set; }
-  public decimal Lag { get; set; }
+  public double Lag { get; set; }
 }
